fix bug on the setting of the final_pos in the individual class

+++ EvolutionVREP/experiments/neuroEvolution/basicEA.cpp
@@ -145,7 +145,7 @@ bool BasicEA::update(const Environment::Ptr & env){
     return true;
 }
 
-void BasicEA::setFitness(size_t indIdx, double fitness){
+void BasicEA::setFitness(size_t indIdx, float fitness){
     currentIndIndex = indIdx;
     population[indIdx]->setFitness(fitness);
 }
